finished form submission and improved validation for profile module
